+++ DonutParade/src/App.jsx
@@ -13,6 +13,7 @@ import ErrorPage from './pages/ErrorPage'
 import AdminOrders from './pages/AdminOrders'
 import AdminEdit from './pages/AdminEdit'
 import AdminHome from './pages/AdminHome'
+import { CartProvider } from './components/CartProvider'
 
 
 
@@ -77,8 +78,9 @@ function App() {
 
   return (
       <div className='App'>
-
+        <CartProvider>
         <RouterProvider router={router} />
+        </CartProvider>
         
       </div>
   )
