+++ DonutParade/src/components/AdminHeader.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from './Navbar'
+//import Navbar from './Navbar'
 import DonutParadeLogo from '../assets/DonutParadeLogoblack.png'
 import '../styles/AdminHeader.css'
