+++ DonutParade/src/components/CartContext.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const CartContext = React.createContext();
