+++ DonutParade/src/pages/Cart.jsx
@@ -1,19 +1,31 @@
-import React from 'react'
+import React, {useContext} from 'react'
 import Header from '../components/Header'
 import Footer from '../components/Footer'
 import '../styles/Cart.css'
+import { CartContext } from '../components/CartContext'
 
 let num = 0;
 let price = 0.00;
-
+// const {cart } = useContext(CartContext);
+// const cartContent = document.querySelector('.cartContent');
+// const yeah = "yeah";
+// cartContent.innerHTML('beforeBegin', `<p>${yeah}</p>`);
 function Cart() {
+
+  
+  
+  
+  
+
   return (
     <div className='Cart'>
 
 
       <Header />
+      
+      
 
-      <section id='cartContent'>
+      <section class='cartContent'>
 
         <h1>Cart</h1>
